Rename parse to decode
