+++ src/index.ts
@@ -1,4 +1,5 @@
 export { getPath, setPath } from "./path.js";
-export { parse, parseWithSchema } from "./parse.js";
+export { decode } from "./decode.js";
+export { parseWithSchema } from "./parse.js";
 export { serialize } from "./serialize.js";
 export type { PathsFromObject, Submission } from "./types.js";
